+++ packages/stream_feeds/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Next release
 - Update API client with renaming `addReaction` to `addActivityReaction` and `deleteReaction` to `deleteActivityReaction`.
+- Update `activity.currentFeed` capabilities when adding or updating activity from websocket events.
 
 ## 0.3.0
 - [BREAKING] Renamed `AppLifecycleStateProvider` to `LifecycleStateProvider` and `AppLifecycleState` to `LifecycleState`.